Don't use libintl.so SONAME for libfbintl.so
POSIX build by default uses libintl.so.1 SONAME for libfbintl.so.1.
This means that on Linux ldconfig will make libintl.so.1 symlink to
libfbintl.so.1.
However, actual libintl is a very different library, incompatible with
libfbintl.
This breaks other software which tries to link to libintl.

Fix it by making libfbintl SONAME match its actual name.