Resolves: tdf#125284 config entries not substituted under Turkish locale
https://bugs.documentfoundation.org/show_bug.cgi?id=125284

This leads to the inability of LibreOffice to create embedded firebird
databases under a Turkish locale.

strcasecmp is used with LC_CTYPE honoured, and...

"These functions shall behave as if the strings had been converted to
lowercase and then a byte comparison performed".

Turkish has the infamous problem that the lowercase of I is the
dotless ı not i.

so strcasecmp("i", "I") != 0 means dir_FOO doesn't caseless match DIR_FOO

firebird/databases.conf contains $(dir_sampleDb) and the string it's compared
to is DIR_SAMPLEDB so a simple fix is to changed DIR_SAMPLEDB to dir_sampledb
and the defalt config will then substitute fine.