Add reuse buffer flag to Sort
Fix crash when accessing already cleared memory. In the following case:
- specific record size;
- the maximum size of the allocated map buffer is greater than MAX_SORT_BUFFER_SIZE;
- unable to allocate memory in dbb_permament pool, throw BadAlloc exception;
- buffer of size MAX_SORT_BUFFER_SIZE must be allocated in the owner's pool.
In the Sort::releaseBuffer procedure, the allocated buffer goes into
the dbb_sort_buffers cache.
After the request is completed, the owner's pool is cleared and
the pointer in dbb_sort_buffers cache will be corrupted. The next
request, which will take the buffer from cache, throw with error.